+++ src/main/java/com/jeju/nanaland/domain/report/service/ReportService.java
@@ -2,7 +2,6 @@ package com.jeju.nanaland.domain.report.service;
 
 import static com.jeju.nanaland.global.exception.ErrorCode.ALREADY_REPORTED;
 import static com.jeju.nanaland.global.exception.ErrorCode.IMAGE_BAD_REQUEST;
-import static com.jeju.nanaland.global.exception.ErrorCode.MAIL_FAIL_ERROR;
 import static com.jeju.nanaland.global.exception.ErrorCode.MEMBER_NOT_FOUND;
 import static com.jeju.nanaland.global.exception.ErrorCode.NANA_INFO_FIX_FORBIDDEN;
 import static com.jeju.nanaland.global.exception.ErrorCode.NOT_FOUND_EXCEPTION;
@@ -15,6 +14,7 @@ import com.jeju.nanaland.domain.common.dto.CompositeDto;
 import com.jeju.nanaland.domain.common.entity.ImageFile;
 import com.jeju.nanaland.domain.common.entity.VideoFile;
 import com.jeju.nanaland.domain.common.service.ImageFileService;
+import com.jeju.nanaland.domain.common.service.MailService;
 import com.jeju.nanaland.domain.common.service.VideoFileService;
 import com.jeju.nanaland.domain.experience.repository.ExperienceRepository;
 import com.jeju.nanaland.domain.festival.repository.FestivalRepository;
@@ -43,12 +43,6 @@ import com.jeju.nanaland.domain.review.entity.Review;
 import com.jeju.nanaland.domain.review.repository.ReviewRepository;
 import com.jeju.nanaland.global.exception.BadRequestException;
 import com.jeju.nanaland.global.exception.NotFoundException;
-import com.jeju.nanaland.global.exception.ServerErrorException;
-import jakarta.mail.Message.RecipientType;
-import jakarta.mail.MessagingException;
-import jakarta.mail.internet.InternetAddress;
-import jakarta.mail.internet.MimeMessage;
-import java.io.UnsupportedEncodingException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -57,13 +51,9 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.env.Environment;
-import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-import org.thymeleaf.context.Context;
-import org.thymeleaf.spring6.SpringTemplateEngine;
 
 @Service
 @RequiredArgsConstructor
@@ -71,8 +61,6 @@ import org.thymeleaf.spring6.SpringTemplateEngine;
 public class ReportService {
 
   private static final int MAX_IMAGE_COUNT = 5;
-  // TODO: 관리자 계정으로 바꾸기
-  private static final String ADMIN_EMAIL = "jyajoo1020@gmail.com";
   private final MemberRepository memberRepository;
   private final ClaimReportVideoFileRepository claimReportVideoFileRepository;
   private final ClaimReportRepository claimReportRepository;
@@ -87,9 +75,8 @@ public class ReportService {
   private final RestaurantRepository restaurantRepository;
   private final ImageFileService imageFileService;
   private final VideoFileService videoFileService;
-  private final Environment env;
-  private final JavaMailSender javaMailSender;
-  private final SpringTemplateEngine templateEngine;
+  private final MailService mailService;
+
   @Value("${cloud.aws.s3.infoFixReportImageDirectory}")
   private String INFO_FIX_REPORT_IMAGE_DIRECTORY;
   @Value("${cloud.aws.s3.claimReportFileDirectory}")
@@ -134,7 +121,7 @@ public class ReportService {
         INFO_FIX_REPORT_IMAGE_DIRECTORY);
 
     // 이메일 전송
-    sendEmailReport(memberInfoDto.getMember().getEmail(), infoFixReport, imageUrls);
+    mailService.sendEmailReport(memberInfoDto.getMember().getEmail(), infoFixReport, imageUrls);
   }
 
   /**
@@ -218,7 +205,7 @@ public class ReportService {
     // 이메일 전송
     List<String> combinedUrls = new ArrayList<>(imageUrls);
     combinedUrls.addAll(videoUrls);
-    sendEmailReport(reqDto.getEmail(), claimReport, combinedUrls);
+    mailService.sendEmailReport(reqDto.getEmail(), claimReport, combinedUrls);
   }
 
   /**
@@ -445,96 +432,5 @@ public class ReportService {
             .claimReport(claimReport)
             .build())
         .collect(Collectors.toList());
-  }
-
-  /**
-   * 메일 전송
-   *
-   * @param memberEmail 회원 이메일
-   * @param report      요청 (InfoFixReport, ClaimReport)
-   * @param urls        파일 URL 리스트
-   * @throws ServerErrorException 메일 전송이 실패한 경우
-   */
-  private void sendEmailReport(String memberEmail, Object report, List<String> urls) {
-    try {
-      MimeMessage mimeMessage = createReportMail(memberEmail, report, urls);
-      javaMailSender.send(mimeMessage);
-    } catch (Exception e) {
-      log.error(e.getMessage());
-      throw new ServerErrorException(MAIL_FAIL_ERROR.getMessage());
-    }
-  }
-
-  /**
-   * 메일 생성
-   *
-   * @param memberEmail 회원 이메일
-   * @param report      요청 (InfoFixReport, ClaimReport)
-   * @param urls        파일 URL 리스트
-   * @return MimeMessage
-   * @throws MessagingException           메일 관련 오류가 발생한 경우
-   * @throws UnsupportedEncodingException 인코딩 오류가 발생한 경우
-   */
-  private MimeMessage createReportMail(String memberEmail, Object report, List<String> urls)
-      throws MessagingException, UnsupportedEncodingException {
-    MimeMessage message = javaMailSender.createMimeMessage();
-    String senderEmail = env.getProperty("spring.mail.username");
-    message.setFrom(new InternetAddress(senderEmail, "Jeju in Nanaland"));
-    message.setRecipients(RecipientType.TO, ADMIN_EMAIL);
-
-    Context context = new Context();
-    String templateName = "";
-    if (report instanceof InfoFixReport infoFixReport) {
-      setInfoFixReportContext(message, context, infoFixReport);
-      templateName = "info-fix-report";
-    } else if (report instanceof ClaimReport claimReport) {
-      setClaimReportContext(memberEmail, message, context, claimReport);
-      templateName = "claim-report";
-    }
-
-    for (int i = 0; i < urls.size(); i++) {
-      context.setVariable("image_" + i, urls.get(i));
-    }
-    message.setText(templateEngine.process(templateName, context), "utf-8", "html");
-
-    return message;
-  }
-
-  /**
-   * 신고 요청 메일 내용 구성
-   *
-   * @param memberEmail 회원 이메일
-   * @param message     내용
-   * @param context     context
-   * @param claimReport ClaimReport
-   * @throws MessagingException 메일 관련 오류가 발생한 경우
-   */
-  private void setClaimReportContext(String memberEmail, MimeMessage message, Context context,
-      ClaimReport claimReport) throws MessagingException {
-    message.setSubject("[Nanaland] 리뷰 신고 요청입니다.");
-    context.setVariable("report_type", claimReport.getReportType());
-    context.setVariable("claim_type", claimReport.getClaimType());
-    context.setVariable("id", claimReport.getId());
-    context.setVariable("content", claimReport.getContent());
-    context.setVariable("email", memberEmail);
-  }
-
-  /**
-   * 정보 수정 제안 요청 메일 내용 구성
-   *
-   * @param message       내용
-   * @param context       context
-   * @param infoFixReport InfoFixReport
-   * @throws MessagingException 메일 관련 오류가 발생한 경우
-   */
-  private void setInfoFixReportContext(MimeMessage message, Context context,
-      InfoFixReport infoFixReport) throws MessagingException {
-    message.setSubject("[Nanaland] 정보 수정 요청입니다.");
-    context.setVariable("fix_type", infoFixReport.getFixType());
-    context.setVariable("category", infoFixReport.getCategory());
-    context.setVariable("language", infoFixReport.getLocale().name());
-    context.setVariable("title", infoFixReport.getTitle());
-    context.setVariable("content", infoFixReport.getContent());
-    context.setVariable("email", infoFixReport.getEmail());
   }
 }
